+++ lang/en/settings/messages.php
@@ -23,4 +23,5 @@ return [
     ],
     'edit_button' => 'Edit settings',
     'private_key_can_not_be_shown' => 'Private key is not shown by security reasons.',
+    'public_key_can_not_be_created' => 'Public key can not be created from the private key. Please, ensure that private key is valid.',
 ];